CHANGES 06-14-2022
Textbox widget is mostly functional now. Just need to add color support for it. Shouldn't be that hard to do so as the textbox byte buffer already makes space for color sequences so I just need to add code in to modify that space whenever the user wants colored text.