CHANGES 07-02-2022
Added Kernel32-based renderer as well as the option to use it on startup.

Widgets that utilize their own ScreenBuffer will now use a standardized CharacterInfo struct to store the ScreenBuffer data.

Because of that, the Kernel32 renderer is the only fully-functional renderer at this time.

The System.Console-based renderer is functional but the TextBox does not draw its buffer properly.

A rework will be done soon to fix the System.Console renderer.